+++ otserv/trunk/container.cpp
@@ -586,7 +586,7 @@ void Container::__removeThing(Thing* thing, uint32_t count)
 	}
 
 	if(item->isStackable() && count != item->getItemCount()){
-		int newCount = std::max(0, (int)(item->getItemCount() - count));
+		uint8_t newCount = (uint8_t)std::max((int32_t)0, (int32_t)(item->getItemCount() - count));
 
 		const double old_weight = -item->getWeight();
 		item->setItemCount(newCount);
