+++ otserv/trunk/item.h
@@ -141,29 +141,29 @@ public:
 	const std::string& getWriter() const {return getStrAttr(ATTR_ITEM_WRITTENBY);}
 
 	void setActionId(uint16_t n) {if(n < 100) n = 100; setIntAttr(ATTR_ITEM_ACTIONID, n);}
-	uint16_t getActionId() const {return getIntAttr(ATTR_ITEM_ACTIONID);}
+	uint16_t getActionId() const {return (uint16_t)getIntAttr(ATTR_ITEM_ACTIONID);}
 
 	void setUniqueId(uint16_t n) {if(n < 1000) n = 1000; setIntAttr(ATTR_ITEM_UNIQUEID, n);}
-	uint16_t getUniqueId() const {return getIntAttr(ATTR_ITEM_UNIQUEID);}
+	uint16_t getUniqueId() const {return (uint16_t)getIntAttr(ATTR_ITEM_UNIQUEID);}
 
 	void setCharges(uint16_t n) {setIntAttr(ATTR_ITEM_CHARGES, n);}
-	uint16_t getCharges() const {return getIntAttr(ATTR_ITEM_CHARGES);}
+	uint16_t getCharges() const {return (uint16_t)getIntAttr(ATTR_ITEM_CHARGES);}
 
 	void setFluidType(uint16_t n) {setIntAttr(ATTR_ITEM_FLUIDTYPE, n);}
-	uint16_t getFluidType() const {return getIntAttr(ATTR_ITEM_FLUIDTYPE);}
+	uint16_t getFluidType() const {return (uint16_t)getIntAttr(ATTR_ITEM_FLUIDTYPE);}
 
 	void setOwner(uint32_t _owner) {setIntAttr(ATTR_ITEM_OWNER, _owner);}
-	uint32_t getOwner() const {return getIntAttr(ATTR_ITEM_OWNER);}
+	uint32_t getOwner() const {return (uint32_t)getIntAttr(ATTR_ITEM_OWNER);}
 
 	void setCorpseOwner(uint32_t _corpseOwner) {setIntAttr(ATTR_ITEM_CORPSEOWNER, _corpseOwner);}
-	uint32_t getCorpseOwner() {return getIntAttr(ATTR_ITEM_CORPSEOWNER);}
+	uint32_t getCorpseOwner() {return (uint32_t)getIntAttr(ATTR_ITEM_CORPSEOWNER);}
 
 	void setDuration(int32_t time) {setIntAttr(ATTR_ITEM_DURATION, time);}
 	void decreaseDuration(int32_t time) {increaseIntAttr(ATTR_ITEM_DURATION, -time);}
-	int32_t getDuration() const {return getIntAttr(ATTR_ITEM_DURATION);}
+	int32_t getDuration() const {return (int32_t)getIntAttr(ATTR_ITEM_DURATION);}
 
 	void setDecaying(ItemDecayState_t decayState) {setIntAttr(ATTR_ITEM_DECAYING, decayState);}
-	uint32_t getDecaying() const {return getIntAttr(ATTR_ITEM_DECAYING);}
+	uint32_t getDecaying() const {return (uint32_t)getIntAttr(ATTR_ITEM_DECAYING);}
 
 protected:
 	enum itemAttrTypes{
@@ -314,7 +314,7 @@ public:
 
 	bool isReadable() const {return items[id].canReadText;}
 	bool canWriteText() const {return items[id].canWriteText;}
-	int32_t getMaxWriteLength() const {return items[id].maxTextLen;}
+	uint16_t getMaxWriteLength() const {return items[id].maxTextLen;}
 
 	int getWorth() const;
 	void getLight(LightInfo& lightInfo);
@@ -351,7 +351,7 @@ public:
 
 	// get the number of items
 	uint16_t getItemCount() const {return count;}
-	void setItemCount(uint16_t n) {count = n;}
+	void setItemCount(uint8_t n) {count = n;}
 
 	static uint32_t countByType(const Item* i, int checkType, bool multiCount);
