+++ otserv/trunk/tile.cpp
@@ -981,7 +981,7 @@ void Tile::__removeThing(Thing* thing, uint32_t count)
 			for (iit = downItems.begin(); iit != downItems.end(); ++iit){
 				if(*iit == item){
 					if(item->isStackable() && count != item->getItemCount()){
-						int newCount = std::max(0, (int)(item->getItemCount() - count));
+						uint8_t newCount = (uint8_t)std::max((int32_t)0, (int32_t)(item->getItemCount() - count));
 						item->setItemCount(newCount);
 
 						const ItemType& it = Item::items[item->getID()];
