+++ otserv/trunk/waitlist.cpp
@@ -122,7 +122,7 @@ bool WaitingList::clientLogin(const Player* player)
 	}
 	else{
 		waitList.push_back(wait);
-		slot = waitList.size();
+		slot = (uint32_t)waitList.size();
 	}
 
 	wait->name = player->getName();
